+++ src/main/java/lotto/Application.java
@@ -2,6 +2,7 @@ package lotto;
 
 import lotto.controller.LottoGameController;
 import lotto.controller.PurchaseLottosRequest;
+import lotto.controller.RetrieveStatisticsRequest;
 import lotto.domain.Lottos;
 import lotto.view.InputView;
 import lotto.view.OutputView;
@@ -11,6 +12,6 @@ public class Application {
         LottoGameController lottoGameController = new LottoGameController();
         Lottos lottos = lottoGameController.purchaseLottos(PurchaseLottosRequest.from(InputView.readPurchasingAmount()));
         OutputView.printPurchasedLottoInformation(lottos);
-        OutputView.printStatisticsForLottos(lottoGameController.retrieveStatistics(lottos, InputView.readWinningLottoNumbers()));
+        OutputView.printStatisticsForLottos(lottoGameController.retrieveStatistics(RetrieveStatisticsRequest.of(lottos, InputView.readWinningLottoNumbers())));
     }
 }
